% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tldr.R
\name{tldr}
\alias{tldr}
\title{Get tldr summary of a system command}
\usage{
tldr(command, platform = get_platform())
}
\arguments{
\item{command}{character. The system command that interests you.}

\item{platform}{character. If the man file is not platform-agnostic,
  search platform specific options. Defaults to your system platform.}
}
\description{
Get tldr summary of a system command
}

